// src/shared/container/providers/PaymentGatewayProvider/interfaces/IAsaasCustomerResponse.ts
export interface IAsaasCustomerResponse {
  id: string;
  object: string;
  dateCreated: string;
  name: string;
  email: string;
  phone: string;
  address: string;
  addressNumber: string;
  complement: string;
  province: string;
  city: string;
  state: string;
  country: string;
  postalCode: string;
  cpfCnpj: string;
  personType: string;
}